+++ Sources/TuistKit/Services/Lint/LintProjectService.swift
@@ -55,7 +55,7 @@ final class LintProjectService {
         logger.notice("Loading the dependency graph")
         if manifests.contains(.workspace) {
             logger.notice("Loading workspace at \(path.pathString)")
-            (graph, _) = try graphLoader.loadWorkspace(path: path)
+            graph = try graphLoader.loadWorkspace(path: path)
         } else if manifests.contains(.project) {
             logger.notice("Loading project at \(path.pathString)")
             (graph, _) = try graphLoader.loadProject(path: path)