+++ Tests/TuistKitTests/Services/Lint/LintCodeServiceTests.swift
@@ -59,7 +59,7 @@ final class LintCodeServiceTests: TuistUnitTestCase {
             "/path1": [.test(target: target01), .test(target: target02), .test(target: target03)],
         ])
         let fakeNoExistTargetName = "Target_999"
-        graphLoader.loadWorkspaceStub = { _ in (graph, Workspace.test()) }
+        graphLoader.loadWorkspaceStub = { _ in graph }
 
         // When
         XCTAssertThrowsSpecific(try subject.run(path: path.pathString, targetName: fakeNoExistTargetName), LintCodeServiceError.targetNotFound(fakeNoExistTargetName))
@@ -78,7 +78,7 @@ final class LintCodeServiceTests: TuistUnitTestCase {
                 "/path1": [.test(target: target01), .test(target: target02), .test(target: target03)],
             ]
         )
-        graphLoader.loadWorkspaceStub = { _ in (graph, Workspace.test()) }
+        graphLoader.loadWorkspaceStub = { _ in graph }
 
         // When
         XCTAssertThrowsSpecific(try subject.run(path: path.pathString, targetName: target01.name), LintCodeServiceError.lintableFilesForTargetNotFound(target01.name))
@@ -118,7 +118,7 @@ final class LintCodeServiceTests: TuistUnitTestCase {
                 "/path1": [.test(target: target01), .test(target: target02), .test(target: target03)],
             ]
         )
-        graphLoader.loadWorkspaceStub = { _ in (graph, Workspace.test()) }
+        graphLoader.loadWorkspaceStub = { _ in graph }
 
         // When
         try subject.run(path: path.pathString, targetName: nil)
@@ -213,7 +213,7 @@ final class LintCodeServiceTests: TuistUnitTestCase {
         let graph = Graph.test(targets: [
             "/path1": [.test(target: target01), .test(target: target02), .test(target: target03)],
         ])
-        graphLoader.loadWorkspaceStub = { _ in (graph, Workspace.test()) }
+        graphLoader.loadWorkspaceStub = { _ in graph }
 
         // When
         try subject.run(path: path.pathString, targetName: target01.name)